refactor(backend): move history API endpoints into farm/history
Move ProjectHistoryListView/RestoreView and GlobalHistoryListView/
RestoreView from farm/views.py into farm/history/views.py, and the
CultureHistoryEntry/CultureRestore serializers into
farm/history/serializers.py. Route table unchanged.
